+++ TutoringSystem/TutoringSystem.Domain/Repositories/IAddressRepository.cs
@@ -6,7 +6,7 @@ namespace TutoringSystem.Domain.Repositories
     public interface IAddressRepository
     {
         Task<bool> AddAddressAsync(Address address);
-        Task<Address> GetAddressByIdAsync(int addressId);
+        Task<Address> GetAddressByIdAsync(long addressId);
         Task<bool> UpdateAddressAsync(Address updatedAddress);
         Task<bool> DeleteAddressAsync(Address address);
     }
