+++ TutoringSystem/TutoringSystem.Domain/Repositories/ISchoolRepository.cs
@@ -6,7 +6,7 @@ namespace TutoringSystem.Domain.Repositories
     public interface ISchoolRepository
     {
         Task<bool> AddSchoolAsync(School school);
-        Task<School> GetSchoolByIdAsync(int schoolId);
+        Task<School> GetSchoolByIdAsync(long schoolId);
         Task<bool> UpdateSchoolAsync(School updatedSchool);
         Task<bool> DeleteSchoolAsync(School school);
     }
